Fix error from previous commit
The code was written in a way that made adding it to a function or
adding another div to the body break its current functionality.
This was fixed by making the column select an element inside the
container div.
This means potentially this function could be used to make multiple
grids inside a single webpage
- Also variable names inside createGrid() was changed to be more
readable